style: fix google-* checks unmasked by the .clang-tidy glob correction
The original .clang-tidy had an inline comment INSIDE the YAML folded scalar
(`google-*      # Google style checks`), which silently malformed the glob and
under-enabled google-build-using-namespace, google-runtime-int and
google-readability-casting. The U4 rewrite parses google-* cleanly, so 38 real
findings surfaced — fix them all (function-local using-declarations replacing
using-directives in the config formatters / tests; long -> int64_t in
proc-system-stats; C-style cast -> static_cast in gstreamer). Also fix a
pre-existing clang-diagnostic-c++11-narrowing in materialize_frame.test.cpp
(uint32 stride from a size_t product) that gcc only warned on but clang-tidy
errors on — masked earlier by grepping only 'warning:'.

Whole tree now exits 0 under the full enforced set (WarningsAsErrors:'*').
Build green; ctest 227/231 (4 hardware tests excluded as in CI).

+++ tests/buffer/materialize_frame.test.cpp
@@ -53,7 +53,7 @@ auto MakeBgr8Frame(Dimensions dims, Bgr color) -> sst::capture::Frame {
     frame.geometry = {dims.width, dims.height};
     frame.captured_at = sst::common::Timestamp{kCaptureMs};
     frame.planes.push_back(sst::capture::FramePlane{
-        .stride = dims.width * kBgrChannels,
+        .stride = static_cast<std::uint32_t>(dims.width * kBgrChannels),
         .data = buf->data(),
         .size = buf->size(),
     });
